Refactor list view to handle new item POSTs

+++ lists/views.py
@@ -27,11 +27,14 @@ def add_item(request, id):
 
 def view_list(request, id):
     list_ = List.objects.get(id=id)
-    item = Item(text=request.POST["item_text"], list=list_)
+    if request.method == 'POST':
+        item = Item(text=request.POST["item_text"], list=list_)
+        return redirect(reverse("view_list"))
+    return render(request, "list.html", {"list": list_})
+
     try: 
         item.full_clean()
         item.save()
     except ValidationError:
         list_.delete()
         return render(request, "home.html", {"error": "You cannot have an empty list item"}) 
-    return render(request, "list.html", {"list": list_})
